pyproject.toml
[tool.poetry]
name = "envy"
version = "0.1.0-dev"
description = "Envy allows developers to simply create project-based development environments that your colleagues will envy."
authors = ["jerluc <me@jerluc.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
blessed = "^1.17.12"
pydantic = "^1.7.3"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
nv = "envy.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
